result = eval("$canvas.width"); 	check_eq(result.value.longy, 300); 	OK !!
    result = eval("$canvas.style");
    ⚠️ Node/string needs to be written to wasm instance and then read back from there

+++ source/Code.h
@@ -39,7 +39,7 @@ class Code;
 //#include <future>
 //std::future<int64> run_wasm(bytes wasm_bytes, int len); // <thread> is not supported on this single threaded system
 #if MY_WASM
-extern "C" int run_wasm(bytes wasm_bytes, int len);
+extern "C" int64 run_wasm(bytes wasm_bytes, int len);
 #else
 extern "C" int64 run_wasm(bytes wasm_bytes, int len);
 #endif
